# openapi.yaml
openapi: 3.0.0
info:
  title: Book Library API
  description: A RESTful API for managing a book library and its authors.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local development server

paths:
  /books:
    get:
      summary: List all books
      tags:
        - Books
      parameters:
        - name: genre
          in: query
          schema:
            type: string
          description: Filter by genre
        - name: author
          in: query
          schema:
            type: string
          description: Filter by author name (substring match across all authors)
      responses:
        '200':
          description: A list of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
    post:
      summary: Create a new book
      tags:
        - Books
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookRequest'
      responses:
        '201':
          description: Book created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate ISBN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /books/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      summary: Get a book by ID
      tags:
        - Books
      responses:
        '200':
          description: Book found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update an existing book
      tags:
        - Books
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookRequest'
      responses:
        '200':
          description: Book updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate ISBN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a book
      tags:
        - Books
      responses:
        '204':
          description: Book deleted
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /authors:
    get:
      summary: List all authors
      tags:
        - Authors
      responses:
        '200':
          description: A list of authors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Author'
    post:
      summary: Create a new author
      tags:
        - Authors
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorRequest'
      responses:
        '201':
          description: Author created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /authors/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      summary: Get an author by ID
      tags:
        - Authors
      responses:
        '200':
          description: Author found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '404':
          description: Author not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /authors/{id}/books:
    get:
      summary: Get all books by a specific author
      tags:
        - Authors
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: A list of books by the author
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        '404':
          description: Author not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Book:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        isbn:
          type: string
        publishedYear:
          type: integer
        genre:
          type: string
    
    BookRequest:
      type: object
      required:
        - title
        - authors
        - isbn
        - publishedYear
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 1
        authors:
          type: array
          items:
            type: string
            maxLength: 255
          minItems: 1
        isbn:
          type: string
          minLength: 1
          description: "Unique ISBN identifier; must be unique across all books"
        publishedYear:
          type: integer
        genre:
          type: string
          maxLength: 255
          description: "Optional genre classification for the book"

    Author:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        nationality:
          type: string
        birthYear:
          type: integer
        books:
          type: array
          items:
            $ref: '#/components/schemas/Book'

    AuthorRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 255
        nationality:
          type: string
          maxLength: 100
        birthYear:
          type: integer

    ErrorResponse:
      type: object
      required:
        - error
        - message
        - status
      properties:
        error:
          type: string
        message:
          type: string
        status:
          type: integer